\name{download_testbam}
\alias{download_testbam}
\title{download_testbam, download bam files of mouse testis and heart}
\usage{
	download_testbam()	
}
\arguments{
    }
\description{
	download bam files of mouse testis and heart
}
\value{
    The function \code{download_testbam} download test data bam files.
}
\author{Ruijia Wang}
\examples{
	download_testbam()
}